requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
psycopg2-binary==2.9.3
redis==4.3.4
pydantic==1.8.2
